requirements.txt
streamlit==1.31.0
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.19
PyPDF2==3.0.1
sentence-transformers==2.3.1
transformers>=4.35.0
torch>=2.0.0
python-dotenv==1.0.1

PyMuPDF>=1.22.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=9.5.0
